refactor: Split hotkey implementation into plugin-hotkey.cpp
